// web/src/lib/generated/ClientMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ClientMessage = { "type": "move", uci: string, } | { "type": "resign" } | { "type": "offer_draw" } | { "type": "accept_draw" } | { "type": "decline_draw" } | { "type": "ping" };