// web/src/lib/generated/GameEnd.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GameOverReason } from "./GameOverReason";
import type { GameResult } from "./GameResult";

/**
 * How a game ended.
 */
export type GameEnd = { result: GameResult, reason: GameOverReason, };